[cudaaligner] Move ukkonen_cpu impl to .cpp file
